+++ tests/integration/execfunc/OutputPage.php
@@ -4,7 +4,7 @@ class OutputPage {
 	public $mBodytext = '';
 
 	public function addHTML( $text ) {
-		$this->mBodytext .= $text; // TODO: This line should be in caused-by
+		$this->mBodytext .= $text;
 	}
 
 	public function getHTML() {