+++ tests/integration/indirectecho/test.php
@@ -4,7 +4,7 @@ class Foo {
 	private $hold = '';
 
 	public function appendHold( $param ) {
-		$this->hold .= $param; // TODO: This line should be in caused-by
+		$this->hold .= $param;
 	}
 
 	public function echoHold() {