+++ Changelog.md
@@ -3,6 +3,8 @@
 ## UNRELEASED
 ### New features
 * Improved taint data for $_FILES
+* Infer when a parameter is passed through by a function (even partially or conditionally), at least in non-complicated cases
+* Improved caused-by lines for setters and some functions that pass their parameters through
 
 ## v3.2.1
 ### Bug fixes